Create GreyTheme YT
his lets you change the Youtube colors to grey and any color of choice
Btw in case you don't know how to do it it's not that it doesn't work. It just means your not doing it correctly so I'll tell you what to do. First you obviously need to download the script and after that say for example you don't like the ugly scroll bar and want the scroll bar to be another color like YouTube's old red or smt your going to want to edit it in settings by clicking on the gear icon ! - anyways hope you enjoy it :D